+++ sw.js
@@ -1,5 +1,5 @@
 
-const CACHE="burger-mastery-v0.1.0";
+const CACHE="burger-mastery-v0.2.0";
 const SHELL=["./","./index.html","./css/app.css","./js/app.js","./js/assets.js","./js/data.js","./manifest.webmanifest","./icons/icon.svg"];
 self.addEventListener("install",e=>e.waitUntil(caches.open(CACHE).then(c=>c.addAll(SHELL)).then(()=>self.skipWaiting())));
 self.addEventListener("activate",e=>e.waitUntil(caches.keys().then(keys=>Promise.all(keys.filter(k=>k!==CACHE).map(k=>caches.delete(k)))).then(()=>self.clients.claim())));